requirements.txt
python-telegram-bot==13.4.1
wfapi==0.8.1
requests==2.25.1
